New Tests and Merge Main

+++ lib/marin/src/marin/scaling_laws/isoflop_analysis.py
@@ -731,7 +731,11 @@ def fit_scaling_laws(
                 continue
 
             # Robust quadratic fit in log10(tokens)
-            a, b, c = robust_quad_logx(jnp.array(sub.tokens.values), jnp.array(sub.loss.values))
+            # Use float64 to avoid int32 overflow for token counts > 2^31
+            a, b, c = robust_quad_logx(
+                jnp.array(sub.tokens.values, dtype=jnp.float64),
+                jnp.array(sub.loss.values, dtype=jnp.float64),
+            )
             fit_curves[(lab, C)] = (a, b, c)
 
             if a == 0:
